# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/empresa/{id}": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Get specific enterprise",
                "description": "Get specific enterprise by id",
                "operationId": "0aae62a86426eb3f6c666da4bdae026c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the enterprise",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved specific enterprise",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 11
                                        },
                                        "user_id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "razao_social": {
                                            "type": "string",
                                            "example": "Razão social inc"
                                        },
                                        "nome_fantasia": {
                                            "type": "string",
                                            "example": "Nome fantasia"
                                        },
                                        "cnpj": {
                                            "type": "string",
                                            "example": "66706525623802"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "ativa"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-07-30T20:35:35.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-07-30T20:35:35.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\Models\\Empresa] 10000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Update an enterprise",
                "description": "Update the details of an existing enterprise",
                "operationId": "5efb82a74a0398d20204102e2752899b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the enterprise to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "razao_social",
                                    "cnpj"
                                ],
                                "properties": {
                                    "razao_social": {
                                        "type": "string",
                                        "example": "Updated Enterprise Name"
                                    },
                                    "nome_fantasia": {
                                        "type": "string",
                                        "example": "Updated Enterprise Name Inc"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "ativa"
                                    },
                                    "cnpj": {
                                        "type": "string",
                                        "example": "12345678901234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Enterprise updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Enterprise updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\Models\\Empresa] 10000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to update the enterprise.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to update the enterprise."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Delete an enterprise",
                "description": "Delete an enterprise by id",
                "operationId": "50f15bcaf5304326f11bd8e99c36b57c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the enterprise to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Enterprise deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\Models\\Empresa] 10000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to remove the enterprise.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to remove the enterprise."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}